requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Audio processing
whisperx==3.7.4
torch==2.8.0
torchaudio==2.8.0
faster-whisper==1.0.3

# File and system utilities
psutil==5.9.6
watchdog==3.0.0
huggingface_hub==0.19.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Additional utilities
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
